refactor(app): P6b — automations/admin tables onto reui/data-grid; old grid deleted
Import repointing for runs-table, automations-table, and
styled-sub-data-grid (old fork's component/prop shapes are identical;
only delta: new DataGridContainer dropped its default border — added
the explicit rounded-lg border className, same idiom as the list
pages). Verified no consumer used dnd row-reorder, so no port (PRD 6b's
parity requirement was a stale assumption). Deleted the 4 old
ui/data-grid* files and the unused sub-data-grid demo (zero importers);
grep confirms zero components/ui/data-grid references remain.
